Update on "explicitly error out in comparison ops when the types don't match"
Explicitly error out in comparison/logical ops when the dtypes of the various input/output tensors don't match. See [this comment](https://github.com/pytorch/pytorch/pull/46399#discussion_r505686406) for more details.

fixes #42660


Differential Revision: [D24335982](https://our.internmc.facebook.com/intern/diff/D24335982)

[ghstack-poisoned]

+++ test/test_autograd.py
@@ -4929,7 +4929,7 @@ def add_test(
             # for-loop bodies don't define scopes, so we have to save the variables
             # we want to close over in some way
             def do_test(self, device, dtype=dtype, name=name, self_size=self_size, args=new_args, test_name=test_name,
-                        output_process_fn=output_process_fn, match_scalar_dtype=match_scalar_dtype):
+                        output_process_fn=output_process_fn):
                 def check(name):
                     is_magic_method = name[:2] == '__' and name[-2:] == '__'
                     is_inplace = name[-1] == "_" and not is_magic_method
